+++ CHANGELOG.md
@@ -1,12 +1,25 @@
 # Change Log
 
-All notable changes to the "nicksays" extension will be documented in this file.
+All notable changes to the "Multi-Cursor-Align" extension will be documented in this file.
 
 ## [Unreleased]
 
-- Initial release
+...
 
-## [0.0.3] - 2023-01-30
+## [1.1.0] - 2023-02-28
+
+### Changed
+
+- Rename extension from `nick_says` to `Multi-Cursor-Align`
+
+### Added
+
+- `nicksays.seek_to_next_occurrence`
+  - Seeks all cursors forward on the same line to the next occurrence of the specified string (or to the end of line if not found)
+- `nicksays.expand_to_next_occurrence`
+  - Expand the cursor forward on the same line to the next occurrence of the specified string (or to the end of line if not found).
+
+## [1.0.3] - 2023-01-30
 
 ### Added